requirements.txt
streamlit
openai
loguru
python-dotenv
requests
PyYAML
pydantic
click